added bare min search
