+++ app/models/translation.rb
@@ -15,7 +15,6 @@
 
 class Translation < ApplicationRecord
   include LanguageFilterable
-  include Searchable
 
   belongs_to :resource, polymorphic: true
   belongs_to :resource_content
