+++ lib/search-v2/query/aggregations.rb
