+++ lib/search-v2/query/fields.rb
