+++ lib/search-v2/query/highlight.rb
