+++ lib/search-v2/query/indices.rb
