+++ lib/search-v2/query/query.rb
